Change 352909  : Show message when viewing an offline version of an article.
