+++ app/src/main/res/values-fr/strings.xml
@@ -20,6 +20,7 @@
   <string name="history_item_delete_undo">Annuler</string>
   <string name="page_does_not_exist_error">Cette page n’existe pas.</string>
   <string name="dialog_title_clear_history">Effacer l’historique de navigation</string>
+  <string name="dialog_offline">Mode hors ligne: Vous êtes deconnecté.</string>
   <string name="dialog_message_clear_history">Cela supprimera tout l’historique de navigation, et fermera tous les onglets ouverts actuellement. Êtes-vous sûr ?</string>
   <string name="nav_item_saved_pages">Pages enregistrées</string>
   <string name="share_via">Partager via</string>
